add ed25519
